feat: refactor socket.io logic into controllers
