+++ src/app.ts
@@ -3,7 +3,6 @@ import express from "express";
 import cookieParser from "cookie-parser";
 import cors from "cors";
 import { RequestContext } from "@mikro-orm/mongodb";
-import cookie from "cookie";
 import { authRouter } from "./modules/auth/auth/auth.routes.js";
 import { classesRouter } from "./modules/class/class-module.routes.js";
 import { clientsRouter } from "./modules/client/client-module.routes.js";
@@ -14,12 +13,8 @@ import { routinesRouter } from "./modules/routine/routine-module.routes.js";
 import { trainerRouter } from "./modules/trainer/trainer/trainer.routes.js";
 import { Server } from "socket.io";
 import { createServer } from "http";
-import jwt from "jsonwebtoken";
-import { environment } from "./config/env.config.js";
-import { Message } from "./modules/chat/message.entity.js";
-import { Client } from "./modules/client/client/client.entity.js";
-import { Trainer } from "./modules/trainer/trainer/trainer.entity.js";
 import { messageRouter } from "./modules/chat/message.routes.js";
+import { setupSocket } from "./utils/socket/socket.js";
 
 const PORT = 3000;
 const app = express();
@@ -61,59 +56,7 @@ app.use((_, res) => {
   return res.status(404).send({ message: "Resource not found" });
 });
 
-io.use((socket, next) => {
-  try {
-    const cookies = socket.handshake.headers.cookie;
-    if (!cookies) throw new Error("No token provided");
-
-    const parsedCookies = cookie.parse(cookies);
-    const token = parsedCookies.auth_token;
-
-    if (!token) throw new Error("Unauthorized. No token found.");
-
-    const decoded = jwt.verify(token, environment.session.jwtSecret);
-    socket.data.user = decoded;
-
-    next();
-  } catch (err) {
-    next(new Error("Authentication error"));
-  }
-});
-
-io.on("connection", (socket) => {
-  socket.on("message", async (data) => {
-    const messageData = JSON.parse(data);
-    console.log(`Mensaje de ${socket.data.user.id}:`, messageData);
-
-    RequestContext.create(orm.em, async () => {
-      try {
-        const newMessage = new Message();
-        newMessage.content = messageData.content;
-        newMessage.createdAt = new Date();
-        const id = messageData.receiver;
-        if (messageData.entity === "client") {
-          newMessage.sender = await orm.em.findOneOrFail(Client, {
-            id: socket.data.user.id,
-          });
-          newMessage.receiver = await orm.em.findOneOrFail(Trainer, { id });
-        } else {
-          newMessage.sender = await orm.em.findOneOrFail(Trainer, {
-            id: socket.data.user.id,
-          });
-          newMessage.receiver = await orm.em.findOneOrFail(Client, { id });
-        }
-
-        await orm.em.persistAndFlush(newMessage);
-      } catch (error) {
-        console.error("Error al guardar mensaje:", error);
-      }
-    });
-
-    io.emit("respuesta", messageData);
-  });
-
-  socket.on("disconnect", () => {});
-});
+setupSocket(io);
 
 httpServer.listen(PORT, () => {
   console.log(`Servidor corriendo en http://localhost:${PORT}/`);
